Add TelegramBot project and update related classes
Added a new project "BuddyLanguage.TelegramBot" to the solution for integration with Telegram. This allows the app to interact with users via Telegram's interface.

Also we refactored the `UpdateHandler` method in `TelegramBotUpdatesListener.cs` to improve readability and error handling.

The `UserVoiceCommandHandler` was also updated to use the newly added `BuddyService` to process incoming voice messages from users and provide feedback asynchronously. Additionally, we added error handling to alert users when their voice messages exceed the acceptable length.