Добавлено логирование в HttpClient TelegramBot
This commit introduces the use of `IHttpClientFactory` for creating `TelegramBotClient` instances, replacing the previous direct instantiation. This change allows for better resource use and HTTP performance. The logging level for "Microsoft.AspNetCore" in both app settings has been increased for better troubleshooting. In the `ChatGPTService`, the used model has been changed to `Gpt3_5_Turbo` as it performs better for the ChatGPT tasks.