Add ILanguage interface and improve language support
Introduced a new ILanguage interface in the BuddyLanguage.Domain.Entities that includes properties such as name, ISO 639-1 language code, and a method to check if a text contains the language. This change will help differentiate and manage multiple languages more effectively.

Other changes include a minor adjustment in UserVoiceCommandHandler in BuddyLanguage.TelegramBot where a newline character was added for better readability of recognized messages.

Furthermore, new package references for LangChain were added in BuddyLanguage.ChatGPTService to improve language recognition and translation services.

Lastly, added a comment in BuddyService.cs to create a custom exception that addresses cases when a user message cannot be recognized.