Refactor user routes and parameter names for clarity
Multiple changes were made to improve readability and understandability of the code.
The GET route 'id' was changed to 'get', and 'telegram-id' was changed to 'get_by_telegram_id',
now routing and functionality are more self-explanatory. Also, the parameter 'userId' was renamed to 'id' in the GetUserByTelegramId function for consistency.
Finally, the 'AddWordEntity' function was renamed to 'AddUser' to accurately depict its functionality.