Update xunit versions and remove unnecessary dependencies
This commit updates the versions of xunit for all test projects from 2.6.1 to 2.6.2 and xunit.runner.visualstudio from 2.5.3 to 2.5.4, to ensure using the latest features and bug fixes.

Furthermore, unnecessary dependencies were removed among multiple projects. In particular, the dependency on BuddyLanguage.WebApi was removed from BuddyLanguage.TelegramBot and some services were no longer added in the Program.cs of the BuddyLanguage.TelegramBot, indicating a separation of concerns and reduction in coupling between components.

Lastly, the AZURE_WEBAPP_NAME environment variable updated to “buddylanguageapi” in the GitHub action for publishing to Azure, to match the new web application name.

These changes aim to uphold good practices in terms of dependency management and align with Azure's naming convention.