Remove platform target specification for 'Release'
This commit removes the specific platform target condition for release configuration in the BuddyLanguage.TelegramBot.csproj. This restriction is unnecessary because it forced the bot to be built specifically for the x64 platform thus limiting its deployability to other platforms. Now, it will build for any platform, increasing its flexibility and compatibility.