Update Azure webapp name and remove release condition
The Azure webapp name was updated from BuddyLanguageTG to BuddyLanguageTG2 in the GitHub workflow. This is likely due to a new version of the web application being created or the old one being replaced for various reasons such as testing or production needs. Additionally, the release condition in the Buddy