Add webhook support for Telegram bot
This commit makes a significant upgrade to the Telegram bot functionality. It implements webhook support with the help of new "HandleUpdate", "ConfigureWebhookBackgroundService" and "ValidateTelegramBotFilterAttribute" methods. Besides, a new "BotConfiguration" setup was introduced in the bot initialization process. Lastly, it also defines different services according to the environment, whether it's a development or production mode, leading to a more robust and scalable approach.